Ollama-QnA-Chatbot/requirements.txt
langchain
langchain_community
langchain_openai
langchain_groq
python-dotenv
langchain-google-genai
